pyproject.toml
[project]
name = "fastapi-article-demo"
version = "0.1.0"
description = "Demo FastAPI project accompanying the article 'FastAPI in Action'"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Wallace Espindola", email = "wallace.espindola@gmail.com"}]


dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.10.0",
    "ruff>=0.6.0",
    "pre-commit>=3.7.0",
    "types-python-jose>=3.3.4.9",
    "types-passlib>=1.7.7.13",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN", "N", "PL", "RUF"]
ignore = ["ANN101", "ANN102"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]
[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
packages = ["app", "tests"]
show_error_codes = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
disallow_untyped_defs = true
strict_optional = true
